Add structures for storage inverted files in the disk
- Add Index method for disk
- Add Engine for Search from the disk

+++ DocCore/Document/DocumentIndex/DocumentIndexHashTable.cs
@@ -42,12 +42,12 @@ namespace DocCore
             this.ht[doc.DocID] = doc;
         }
 
-        public Document Search(long docID)
+        public Document Search(int docID)
         {
             return this.ht[docID] as Document;
         }
 
-        public void Delete(long docID)
+        public void Delete(int docID)
         {
             this.ht.Remove(docID);
         }
